Rebased to main after PR#6926 remove "*" imports
